+++ app/articles/permissions.py
@@ -6,10 +6,10 @@ from rest_framework.views import APIView
 
 
 class AuthorOrReadOnly(permissions.BasePermission):
-    def has_permission(self, request: Request, view: APIView) -> Any:
-        return request.user.is_authenticated
+    def has_permission(self, request: Request, view: APIView) -> bool:
+        return bool(request.user.is_authenticated)
 
     def has_object_permission(
         self, request: Request, view: APIView, obj: Any
-    ) -> Any:
-        return obj.author == request.user
+    ) -> bool:
+        return bool(obj.author == request.user)
