+++ app/articles/urls.py
@@ -3,16 +3,10 @@ from django.urls import path
 from app.articles.views import ArticleDetailView, ArticleListCreateView
 
 urlpatterns = [
-    path("all-articles/", ArticleListCreateView.as_view(), name="articles"),
+    path("articles/", ArticleListCreateView.as_view(), name="article-list"),
     path(
-        "article/<slug:slug>/detail/",
+        "article/<slug:slug>/",
         ArticleDetailView.as_view(),
-        name="article-update",
+        name="article-detail",
     ),
-    path(
-        "delete/<slug:slug>/detail/",
-        ArticleDetailView.as_view(),
-        name="article-delete",
-    ),
-    path("create/", ArticleListCreateView.as_view(), name="create"),
 ]
