+++ app/user/tests/test_views.py
@@ -1,12 +1,11 @@
 import json
-from typing import Any, Dict
+from typing import Any
 from unittest.mock import patch
 
 from django.contrib.auth import get_user_model
 from django.contrib.auth.tokens import PasswordResetTokenGenerator
 from django.core import mail
 from django.test import TestCase
-from django.test.client import BOUNDARY, MULTIPART_CONTENT, encode_multipart
 from django.urls import reverse
 from django.utils.encoding import force_bytes
 from django.utils.http import urlsafe_base64_encode
@@ -14,12 +13,10 @@ from faker import Faker
 from rest_framework import status
 from rest_framework.test import APIClient, APITestCase
 
-from app.articles.models import Article
 from app.user.models import Profile
 from app.user.token import account_activation_token
 
-from .mock import sample_image
-from .mocks import test_image, test_user
+from .mocks import test_image
 
 User = get_user_model()
 fake = Faker()
@@ -119,23 +116,29 @@ class TestProfileView(APITestCase):
     Test for Profile Model view
     """
 
-    def setUp(self) -> None:
-        self.user_test = User.objects.create_user(**test_user)
+    @classmethod
+    def setUpClass(cls) -> None:
+        """
+        setup data to use in the tests
+        """
+        super().setUpClass()
+        cls.password = fake.password()  # type: ignore[attr-defined]
+        cls.user = User.objects.create_user(  # type: ignore[attr-defined]
+            username=fake.name(), email=fake.email(), password=cls.password  # type: ignore[attr-defined]
+        )
 
+    def setUp(self) -> None:
         self.client = APIClient()
 
     @property
     def bearer_token(self) -> dict:
         """
-        Authentication function
+        Authentication function: logs in user
         """
         login_url = reverse("login")
         response = self.client.post(
             login_url,
-            data={
-                "email": test_user["email"],
-                "password": test_user["password"],
-            },
+            data={"email": self.user.email, "password": self.password},  # type: ignore[attr-defined]
             format="json",
         )
         token = json.loads(response.content).get("access")
@@ -148,23 +151,18 @@ class TestProfileView(APITestCase):
         """
         Test updating of profile by user
         """
-        Profile.objects.create(user=self.user_test)
+        Profile.objects.create(user=self.user)  # type: ignore[attr-defined]
         data = {
             "bio": "Do I function",
             "image": test_image,
         }
-        url = reverse("profile", kwargs={"user": self.user_test.id})
-        # self.client.defaults[
-        #     "HTTP_AUTHORIZATION"
-        # ] = f"Bearer {self.bearer_token}"
+        url = reverse("profile", kwargs={"user": self.user.id})  # type: ignore[attr-defined]
         response = self.client.patch(
-            url,
-            data=encode_multipart(data=data, boundary=BOUNDARY),
-            content_type=MULTIPART_CONTENT,
-            enctype="multipart/form-data",
-            **self.bearer_token,
+            url, data=data, format="json", **self.bearer_token
         )
+        import pdb
 
+        pdb.set_trace()
         self.assertTrue(upload_resource.called)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
 
@@ -269,99 +267,3 @@ class TestPasswordReset(TestCase):
         resp = self.client.post(reset_url)
         self.assertEqual(resp.status_code, 400)
         self.assertIn("This field is required", str(resp.data))  # type: ignore[attr-defined]
-
-
-class TestArticleViews(TestCase):
-    """
-    Tests for articles app views
-    """
-
-    user: Any
-    article: Any
-    password: str
-    data: Dict[str, Any]
-
-    @classmethod
-    def setUpClass(cls) -> None:
-        """
-        setup data to use in the tests
-        """
-        super().setUpClass()
-        cls.password = fake.password()
-        cls.user = User.objects.create_user(
-            username=fake.name(), email=fake.email(), password=cls.password
-        )
-        cls.article = Article.objects.create(
-            title=fake.name(),
-            description=fake.text(),
-            body=fake.text(),
-            image=sample_image(),
-            author=cls.user,
-        )
-
-        cls.data = {
-            "title": fake.texts(nb_texts=2),
-            "description": fake.paragraph(nb_sentences=3),
-            "body": fake.paragraph(nb_sentences=20),
-            "image": sample_image(),
-            "taglist": f"{fake.word()}, {fake.word()}",
-        }
-
-    @property
-    def bearer_token(self) -> dict:
-        """
-        Authentication function: logs in user
-        """
-        login_url = reverse("login")
-        response = self.client.post(
-            login_url,
-            data={"email": self.user.email, "password": self.password},
-            format="json",
-        )
-        token = json.loads(response.content).get("access")
-        return {"HTTP_AUTHORIZATION": f"Bearer {token}"}
-
-    @patch(
-        "cloudinary.uploader.upload_resource", return_value=fake.image_url()
-    )
-    def test_create_article(self, upload_resource: Any) -> None:
-        """
-        Testing creation of articles
-        """
-        count = Article.objects.count()
-        response = self.client.post(
-            reverse("create"),
-            data=self.data,
-            format="json",
-            **self.bearer_token,
-        )
-        self.assertTrue(upload_resource.called)
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(Article.objects.count(), count + 1)
-
-    def test_delete_article(self) -> None:
-        """
-        Test deletion of article by the author
-        """
-        count = Article.objects.count()
-        response = self.client.delete(
-            reverse("article-delete", kwargs={"slug": self.article.slug}),
-            **self.bearer_token,
-        )
-        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
-        self.assertEqual(Article.objects.count(), count - 1)
-
-    def test_delete_article_without_authentication(self) -> None:
-        """
-        Test deletion of article by a foreign user
-        """
-        count = Article.objects.count()
-        response = self.client.delete(
-            reverse("article-delete", kwargs={"slug": self.article.slug}),
-        )
-        self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
-        self.assertEqual(Article.objects.count(), count)
-        self.assertEqual(
-            json.loads(response.content).get("detail"),
-            "Authentication credentials were not provided.",
-        )
